move contract types out of the root level
