+++ protostar/testing/cheatcodes/prepare_cheatcode.py
@@ -1,4 +1,5 @@
 import collections
+from dataclasses import dataclass
 from typing import Any, Callable, List, Optional
 
 from starkware.starknet.core.os.contract_address.contract_address import (
@@ -12,7 +13,16 @@ from protostar.starknet.data_transformer import (
     from_python_transformer,
     DataTransformerException,
 )
-from protostar.contract_types import DeclaredContract, PreparedContract
+
+from .declare_cheatcode import DeclaredContract
+
+
+@dataclass(frozen=True)
+class PreparedContract:
+    constructor_calldata: List[int]
+    contract_address: int
+    class_hash: int
+    salt: int
 
 
 class PrepareCheatcode(Cheatcode):
